requirements.txt
openai
pandas
numpy
spacy
nltk
scikit-learn
sentence-transformers
matplotlib
streamlit>=1.40.0
plotly
groq
python-dotenv